AnimatedEntity + New Spritesheets
